Apagando classe abstrata e terminando rota de produtos

+++ src/main/java/the/coyote/sales/Model/Entity/Enderecos.java
@@ -14,7 +14,7 @@ public class Enderecos {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE)
-	private Long idEndereco;
+	private int idEndereco;
 	
 	private String rua;
 	private String numero;
